moving fixtures straight into conftest
enables autocompletion and gets rid of a
superfluous folder. also added two
fixtures to make the server tests a little
dry-er.
