+++ panel/tests/widgets/test_base.py
@@ -5,7 +5,7 @@ import pytest
 
 from panel.io import block_comm
 from panel.widgets import CompositeWidget, TextInput, Widget
-from panel._testing.util import check_layoutable_properties
+from panel.tests.util import check_layoutable_properties
 
 all_widgets = [w for w in param.concrete_descendents(Widget).values()
                if not w.__name__.startswith('_') and not issubclass(w, CompositeWidget)]
